refactor: streamline peer discovery and value storage methods in KadDHT
