+++ libp2p/kad_dht/kad_dht.py
@@ -93,30 +93,10 @@ class KadDHT(Service):
             # Periodically refresh the routing table
             await self.refresh_routing_table()
 
-            # Connect to all known peers
-            logger.info("Connecting to all known peers...")
-            await self.connect_to_all_known_peers()
-
-            # Actively trigger peer discovery by querying for our own ID
-            try:
-                logger.info("Triggering peer discovery with find_peer %s",self.local_peer_id)
-                peers = await self.find_peer(self.local_peer_id)
-                logger.info("Peer discovery query completed")
-                if peers:
-                    logger.info("found peers: %s", peers)
-                    for peer in peers:
-                        if peer != self.local_peer_id:
-                            logger.info("peer id: %s", peer.peer_id)
-                            logger.info("peer multiaddress: %s", peer.addrs)
-            except Exception as e:
-                logger.warning(f"Peer discovery query failed: {e}")
-                
             # Check if it's time to republish provider records
             current_time = time.time()
-            if current_time - self._last_provider_republish > PROVIDER_RECORD_REPUBLISH_INTERVAL:
-                logger.info("Republishing provider records...")
-                await self._republish_provider_records()
-                self._last_provider_republish = current_time
+            # await self._republish_provider_records()
+            self._last_provider_republish = current_time
             
             # Clean up expired values and provider records
             self.value_store.cleanup_expired()
@@ -125,249 +105,6 @@ class KadDHT(Service):
             # Wait before next maintenance cycle
             await trio.sleep(ROUTING_TABLE_REFRESH_INTERVAL)
             
-    async def _republish_provider_records(self) -> None:
-        """Republish all provider records for content this node is providing."""
-        for key in self._providing_keys:
-            logger.info(f"Republishing provider record for key {key.hex()}")
-            await self.provide(key)
-    
-    # Content provider methods
-    
-    async def provide(self, key: bytes) -> bool:
-        """
-        Advertise that this node can provide a piece of content.
-        
-        Finds the k closest peers to the key and sends them ADD_PROVIDER messages.
-        
-        Args:
-            key: The content key (multihash) to advertise
-            
-        Returns:
-            bool: True if the advertisement was successful
-        """
-        # Add to local provider store
-        local_addrs = []
-        for addr in self.host.get_addrs():
-            local_addrs.append(addr)
-        
-        local_peer_info = PeerInfo(self.local_peer_id, local_addrs)
-        self.provider_store.add_provider(key, local_peer_info)
-        
-        # Track that we're providing this key
-        self._providing_keys.add(key)
-        
-        # Find the k closest peers to the key
-        closest_peers = await self.peer_routing.find_closest_peers_network(key)
-        logger.info(f"Found {len(closest_peers)} peers close to key {key.hex()} for provider advertisement")
-        
-        # Send ADD_PROVIDER messages to these peers
-        success_count = 0
-        for peer_id in closest_peers:
-            if peer_id == self.local_peer_id:
-                continue
-                
-            try:
-                success = await self._send_add_provider(peer_id, key)
-                if success:
-                    success_count += 1
-            except Exception as e:
-                logger.warning(f"Failed to send ADD_PROVIDER to {peer_id}: {e}")
-                
-        return success_count > 0
-        
-    async def _send_add_provider(self, peer_id: ID, key: bytes) -> bool:
-        """
-        Send ADD_PROVIDER message to a specific peer.
-        
-        Args:
-            peer_id: The peer to send the message to
-            key: The content key being provided
-            
-        Returns:
-            bool: True if the message was successfully sent and acknowledged
-        """
-        try:
-            # Open a stream to the peer
-            stream = await self.host.new_stream(peer_id, [PROTOCOL_ID])
-            
-            try:
-                # Get our addresses to include in the message
-                addrs = []
-                for addr in self.host.get_addrs():
-                    addrs.append(addr.to_bytes())
-                
-                # Create the ADD_PROVIDER message
-                message = Message()
-                message.type = Message.MessageType.ADD_PROVIDER
-                message.key = key
-                
-                # Add our provider info
-                provider = message.providerPeers.add()
-                provider.id = self.local_peer_id.to_bytes()
-                provider.addrs.extend(addrs)
-                
-                # Serialize and send the message
-                proto_bytes = message.SerializeToString()
-                await stream.write(len(proto_bytes).to_bytes(4, byteorder='big'))
-                await stream.write(proto_bytes)
-                
-                # Read response (length prefix)
-                length_bytes = await stream.read(4)
-                if len(length_bytes) < 4:
-                    return False
-                    
-                response_length = int.from_bytes(length_bytes, byteorder='big')
-                
-                # Read response data
-                response_bytes = await stream.read(response_length)
-                if len(response_bytes) < response_length:
-                    return False
-                    
-                # Parse response
-                response = Message()
-                response.ParseFromString(response_bytes)
-                
-                # Check response type
-                return response.type == Message.MessageType.ADD_PROVIDER
-                
-            finally:
-                await stream.close()
-                
-        except Exception as e:
-            logger.warning(f"Error sending ADD_PROVIDER to {peer_id}: {e}")
-            return False
-            
-    async def find_providers(self, key: bytes, count: int = 20) -> List[PeerInfo]:
-        """
-        Find content providers for a given key.
-        
-        Args:
-            key: The content key to look for
-            count: Maximum number of providers to return
-            
-        Returns:
-            List[PeerInfo]: List of content providers
-        """
-        # Check local provider store first
-        local_providers = self.provider_store.get_providers(key)
-        if local_providers:
-            logger.info(f"Found {len(local_providers)} providers locally for {key.hex()}")
-            return local_providers[:count]
-        logger.info("local providers are %s", local_providers)
-        # Find the closest peers to the key
-        closest_peers = await self.peer_routing.find_closest_peers_network(key)
-        logger.info(f"Searching {len(closest_peers)} peers for providers of {key.hex()}")
-        
-        # Query these peers for providers
-        all_providers = []
-        for peer_id in closest_peers:
-            if peer_id == self.local_peer_id:
-                continue
-                
-            try:
-                providers = await self._get_providers_from_peer(peer_id, key)
-                if providers:
-                    # Add providers to our local store
-                    for provider in providers:
-                        self.provider_store.add_provider(key, provider)
-                    
-                    # Add to our result list
-                    all_providers.extend(providers)
-                    
-                    # Stop if we've found enough providers
-                    if len(all_providers) >= count:
-                        break
-            except Exception as e:
-                logger.warning(f"Failed to get providers from {peer_id}: {e}")
-                
-        return all_providers[:count]
-        
-    async def _get_providers_from_peer(self, peer_id: ID, key: bytes) -> List[PeerInfo]:
-        """
-        Get content providers from a specific peer.
-        
-        Args:
-            peer_id: The peer to query
-            key: The content key to look for
-            
-        Returns:
-            List[PeerInfo]: List of provider information
-        """
-        try:
-            # Open a stream to the peer
-            stream = await self.host.new_stream(peer_id, [PROTOCOL_ID])
-            
-            try:
-                # Create the GET_PROVIDERS message
-                message = Message()
-                message.type = Message.MessageType.GET_PROVIDERS
-                message.key = key
-                
-                # Serialize and send the message
-                proto_bytes = message.SerializeToString()
-                await stream.write(len(proto_bytes).to_bytes(4, byteorder='big'))
-                await stream.write(proto_bytes)
-                
-                # Read response (length prefix)
-                length_bytes = b""
-                remaining = 4
-                while remaining > 0:
-                    chunk = await stream.read(remaining)
-                    if not chunk:
-                        return []
-                        
-                    length_bytes += chunk
-                    remaining -= len(chunk)
-                    
-                response_length = int.from_bytes(length_bytes, byteorder='big')
-                
-                # Read response data
-                response_bytes = b""
-                remaining = response_length
-                while remaining > 0:
-                    chunk = await stream.read(remaining)
-                    if not chunk:
-                        return []
-                        
-                    response_bytes += chunk
-                    remaining -= len(chunk)
-                    
-                # Parse response
-                response = Message()
-                response.ParseFromString(response_bytes)
-                
-                # Check response type
-                if response.type != Message.MessageType.GET_PROVIDERS:
-                    return []
-                    
-                # Extract provider information
-                providers = []
-                for provider_proto in response.providerPeers:
-                    try:
-                        # Create peer ID from bytes
-                        provider_id = ID(provider_proto.id)
-                        
-                        # Convert addresses to Multiaddr
-                        addrs = []
-                        for addr_bytes in provider_proto.addrs:
-                            try:
-                                addrs.append(Multiaddr(addr_bytes))
-                            except:
-                                pass  # Skip invalid addresses
-                                
-                        # Create PeerInfo and add to result
-                        providers.append(PeerInfo(provider_id, addrs))
-                    except Exception as e:
-                        logger.warning(f"Failed to parse provider info: {e}")
-                        
-                return providers
-                
-            finally:
-                await stream.close()
-                
-        except Exception as e:
-            logger.warning(f"Error getting providers from {peer_id}: {e}")
-            return []
             
     async def handle_stream(self, stream: INetStream) -> None:
         """
@@ -672,7 +409,6 @@ class KadDHT(Service):
 
     async def refresh_routing_table(self) -> None:
         """Refresh the routing table."""
-        logger.info("Refreshing routing table1...")
         await self.peer_routing.refresh_routing_table()
         
     # Peer routing methods
@@ -687,46 +423,9 @@ class KadDHT(Service):
         Returns:
             Optional[PeerInfo]: The peer information if found, None otherwise
         """
-        try:
-            # First check if the peer is in our peerstore
-            # try:
-            #     addrs = self.host.get_peerstore().addrs(peer_id)
-            #     logger.info(f"address of self peer: {addrs}")
-            #     if addrs:
-            #         logger.debug(f"Found peer {peer_id} in local peerstore")
-            #         return PeerInfo(peer_id, addrs)
-            # except Exception:
-            #     pass
-                
-            # If not, forward to peer_routing with proper error handling
-            
-            return await self.peer_routing.find_peer(peer_id)
-        except Exception as e:
-            logger.warning(f"Error finding peer: {e}")
-            return None
-    
-    async def connect_to_all_known_peers(self):
-        """
-        Attempt to connect to all peers currently in the routing table.
-        """
-        logger.info("Connecting to all known peers method...")
-        peer_ids = self.routing_table.get_peer_ids()
-        for peer_id in peer_ids:
-            if peer_id == self.local_peer_id:
-                continue
-            try:
-                # Try to get PeerInfo from the peerstore
-                peerstore = getattr(self.host, "get_peerstore", None)
-                if peerstore:
-                    addrs = self.host.get_peerstore().addrs(peer_id)
-                    if addrs:
-                        peer_info = PeerInfo(peer_id, addrs)
-                        await self.host.connect(peer_info)
-                        logger.info(f"Connected to discovered peer: {peer_id.pretty()}")
-            except Exception as e:
-                logger.debug(f"Failed to connect to peer {peer_id}: {e}")
-            
         
+        return await self.peer_routing.find_peer(peer_id)
+                  
     async def put_value(self, key: Union[str, bytes], value: bytes) -> None:
         """
         Store a value in the DHT.
@@ -748,7 +447,7 @@ class KadDHT(Service):
         
         # 3. Store at remote peers
         for peer in closest_peers:
-            await self._store_at_peer(peer, key, value)
+            await self.value_store._store_at_peer(peer, key, value)
             
     async def get_value(self, key: str) -> Optional[bytes]:
         """Retrieve a value from the DHT."""
@@ -775,180 +474,6 @@ class KadDHT(Service):
                 
     # Add these methods in the Utility methods section
 
-    async def _store_at_peer(self, peer_id: ID, key: str, value: bytes) -> bool:
-        """
-        Store a value at a specific peer.
-        
-        Args:
-            peer_id: The ID of the peer to store the value at
-            key: The key to store
-            value: The value to store
-            
-        Returns:
-            bool: True if the value was successfully stored, False otherwise
-        """
-        try:
-            # Don't try to store at ourselves
-            if peer_id == self.local_peer_id:
-                return True
-                
-            logger.info(f"Storing value for key {key} at peer {peer_id}")
-            
-            # Open a stream to the peer
-            stream = await self.host.new_stream(peer_id, [PROTOCOL_ID])
-            logger.info(f"Opened stream to peer {peer_id}")
-
-            # Create the PUT_VALUE message with protobuf
-            message = Message()
-            message.type = Message.MessageType.PUT_VALUE
-            
-            # Convert key to bytes if it's a string
-            key_bytes = key if isinstance(key, bytes) else key.encode()
-            
-            # Set message fields
-            message.key = key_bytes
-            message.record.key = key_bytes
-            message.record.value = value
-            message.record.timeReceived = str(time.time())
-            
-            # Serialize and send the protobuf message with length prefix
-            proto_bytes = message.SerializeToString()
-            await stream.write(len(proto_bytes).to_bytes(4, "big"))
-            await stream.write(proto_bytes)
-            logger.info("Sent PUT_VALUE protobuf message")
-
-            # Read response length (4 bytes)
-            length_bytes = await stream.read(4)
-            if len(length_bytes) < 4:
-                logger.warning("Failed to read response length")
-                return False
-                
-            response_length = int.from_bytes(length_bytes, byteorder='big')
-            
-            # Read response
-            response_bytes = await stream.read(response_length)
-            if len(response_bytes) < response_length:
-                logger.warning("Failed to read complete response")
-                return False
-                
-            # Parse protobuf response
-            response = Message()
-            response.ParseFromString(response_bytes)
-            
-            # Check if response is valid
-            if response.type == Message.MessageType.PUT_VALUE:
-                logger.debug(f"Successfully stored value at peer {peer_id}")
-                return True
-                
-            return False
-            
-        except Exception as e:
-            logger.warning(f"Failed to store value at peer {peer_id}: {e}")
-            return False
-            
-        finally:
-            if stream:
-                await stream.close()
-            
-    async def _get_from_peer(self, peer_id: ID, key: str) -> Optional[bytes]:
-        """
-        Retrieve a value from a specific peer.
-        
-        Args:
-            peer_id: The ID of the peer to retrieve the value from
-            key: The key to retrieve
-            
-        Returns:
-            Optional[bytes]: The value if found, None otherwise
-        """
-        stream = None
-        try:
-            # Don't try to get from ourselves
-            if peer_id == self.local_peer_id:
-                return None
-                
-            logger.info(f"Getting value for key {key} from peer {peer_id}")
-            
-            # Open a stream to the peer
-            stream = await self.host.new_stream(peer_id, [PROTOCOL_ID])
-            logger.info(f"Opened stream to peer {peer_id} for GET_VALUE")
-
-            # Create the GET_VALUE message using protobuf
-            message = Message()
-            message.type = Message.MessageType.GET_VALUE
-            
-            # Convert key to bytes if it's a string
-            key_bytes = key if isinstance(key, bytes) else key.encode()
-            message.key = key_bytes
-            
-            # Serialize and send the protobuf message
-            proto_bytes = message.SerializeToString()
-            await stream.write(len(proto_bytes).to_bytes(4, "big"))
-            await stream.write(proto_bytes)
-            
-            # Read response length (4 bytes)
-            length_bytes = b""
-            remaining = 4
-            while remaining > 0:
-                chunk = await stream.read(remaining)
-                if not chunk:
-                    logger.debug(f"Connection closed by peer {peer_id} while reading length")
-                    return None
-                    
-                length_bytes += chunk
-                remaining -= len(chunk)
-                
-            response_length = int.from_bytes(length_bytes, byteorder='big')
-            
-            # Read response data
-            response_bytes = b""
-            remaining = response_length
-            while remaining > 0:
-                chunk = await stream.read(remaining)
-                if not chunk:
-                    logger.debug(f"Connection closed by peer {peer_id} while reading data")
-                    return None
-                    
-                response_bytes += chunk
-                remaining -= len(chunk)
-                
-            # Parse protobuf response
-            try:
-                response = Message()
-                response.ParseFromString(response_bytes)
-                logger.info(f"Received protobuf response from peer {peer_id}, type: {response.type}")
-                
-                # Process protobuf response
-                if (response.type == Message.MessageType.GET_VALUE and
-                    response.HasField('record') and 
-                    response.record.value):
-                    logger.debug(f"Received value for key {key_bytes.hex()} from peer {peer_id}")
-                    return response.record.value
-                    
-            except Exception as proto_err:
-                # Fall back to JSON for backward compatibility
-                logger.warning(f"Failed to parse as protobuf, trying JSON: {proto_err}")
-                try:
-                    json_response = json.loads(response_bytes.decode())
-                    if json_response.get("type") == "VALUE" and "value" in json_response:
-                        value = json_response["value"]
-                        if isinstance(value, str):
-                            value = value.encode()
-                        logger.debug(f"Received JSON value for key {key} from peer {peer_id}")
-                        return value
-                except Exception as json_err:
-                    logger.error(f"Failed to parse as JSON too: {json_err}")
-                
-            return None
-            
-        except Exception as e:
-            logger.warning(f"Failed to get value from peer {peer_id}: {e}")
-            return None
-            
-        finally:
-            if stream:
-                await stream.close()
-        
     # Utility methods
     
     def add_peer(self, peer_id: ID) -> bool:
@@ -961,19 +486,7 @@ class KadDHT(Service):
         Returns:
             bool: True if peer was added or updated, False otherwise
         """
-        try:
-            # Get addresses from the peerstore if available
-            addrs = self.host.get_peerstore().addrs(peer_id)
-            if addrs:
-                # Create PeerInfo object and add to routing table
-                peer_info = PeerInfo(peer_id, addrs)
-                return self.routing_table.add_peer(peer_info)
-            else:
-                logger.warning(f"No addresses found for peer {peer_id}, cannot add to routing table")
-                return False
-        except Exception as e:
-            logger.warning(f"Error adding peer {peer_id} to routing table: {e}")
-            return False
+        return self.routing_table.add_peer(peer_id)
         
     def get_routing_table_size(self) -> int:
         """
